+++ lib/features/item/domain/services/item_service_interface.dart
@@ -26,7 +26,7 @@ abstract class ItemServiceInterface {
   List<int> initializeAddonQtyList(List<AddOns>? addOns);
   Future<String> prepareVariationType(List<ChoiceOptions>? choiceOptions, List<int>? variationIndex);
   int setAddOnQuantity(bool isIncrement, int addOnQty);
-  Future<double> setQuantity(bool isIncrement, bool moduleStock, int? stock, double qty, int? quantityLimit, {bool getxSnackBar = false});
+  Future<double> setQuantity(bool isIncrement, bool moduleStock, double? stock, double qty, int? quantityLimit, {bool getxSnackBar = false});
   List<List<bool?>> setNewCartVariationIndex(int index, int i, List<FoodVariation>? foodVariations, List<List<bool?>> selectedVariations);
   int selectedVariationLength(List<List<bool?>> selectedVariations, int index);
   double? getStartingPrice(Item item);
